requirements.txt
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
numpy
plotly
rasterio
shapely
Pillow
matplotlib
